code integration

+++ matlab/crossvalidateSigma.m
@@ -26,8 +26,8 @@ function bestSigma = crossvalidateSigma(train, test, VC, minValue, maxValue, ste
           clear h;
         end
         
-        bof_train=(cat(1,train.bof));
-        bof_test=(cat(1,test.bof));
+        bof_train=double(cat(1,train.bof));
+        bof_test=double(cat(1,test.bof));
         labels_train=cat(1,train.class);
         labels_test=cat(1,test.class);
